Refactor: Add ShiftGenerator module to store methods creating shift
